feat: organize the modules in separate files and dirs and init repos with git2
