+++ Cargo.toml
@@ -13,6 +13,7 @@ categories = ["command-line-utilities"]
 [dependencies]
 clap = { version = "4.5.29", features = ["derive"] }
 crossterm = "0.28.1"
+git2 = "0.20.0"
 ratatui = "0.29.0"
 
 [dev-dependencies]
